+++ Server/src/mail.c
@@ -6127,7 +6127,7 @@ void myfgets(char *buf, FILE *fpt)
 
   pt1 = buf;
   in = fgetc(fpt);
-  while ((in != EOF) && (in) && (in != '\1')) {
+  while (((int)in != EOF) && (in) && (in != '\1')) {
     *pt1++ = in;
     in = fgetc(fpt);
   }
@@ -6168,7 +6168,7 @@ void mnuke_read()
   nuke1 = fopen(nukename, "r");
   if (nuke1) {
     mudstate.nuke_status = 1;
-    while ((in1 = fgetc(nuke1)) != EOF) {
+    while ((int)(in1 = fgetc(nuke1)) != EOF) {
       *buf1 = in1;
       myfgets(buf1+1,nuke1);
       mail_wipe(GOD,buf1);
@@ -6227,7 +6227,7 @@ void mail_load(dbref player)
     mail_error(player,MERR_DB);
     return;
   }
-  while ((input = fgetc(dump1)) != EOF) {
+  while ((int)(input = fgetc(dump1)) != EOF) {
     input = input - 'A' + 1;
     *(int *)sbuf1 = (int)input;
     keydata.dptr = sbuf1;
@@ -6401,7 +6401,7 @@ void mail_load(dbref player)
     dbm_store(mailfile,keydata,infodata,DBM_REPLACE);
   }
   fclose(dump1);
-  while ((input = fgetc(dump2)) != EOF) {
+  while ((int)(input = fgetc(dump2)) != EOF) {
     input = input - 'A' + 1;
     *(int *)sbuf1 = (int)input;
     keydata.dptr = sbuf1;
